feature: removed name from pokemon variety card
