+++ src/components/feedbacks/PokemonVarietyCardLoadingFeedback.tsx
@@ -1,14 +1,10 @@
-import { CenteredFlexCol, Title } from "../main-components"
 import HoverableGrowthFeedback from "./HoverableGrowthFeedback"
 import RotatingPokeballFeedback from "./RotatingPokeballFeedback"
 
 function PokemonVarietyCardLoadingFeedback({ name }: { name: string }) {
   return ( 
     <HoverableGrowthFeedback>
-      <CenteredFlexCol $gap="1rem"> 
-        <RotatingPokeballFeedback pokemonId={name} width={175} height={175}/>
-        <Title $color="#fff">{name}</Title>
-      </CenteredFlexCol>
+      <RotatingPokeballFeedback pokemonId={name} width={250} height={250}/>
     </HoverableGrowthFeedback>
    )
 }
