+++ src/types/pokemon-related-types.ts
@@ -1,4 +1,4 @@
-type PokemonType = {
+export type PokemonType = {
   type: NamedAPIResource
   slot: number, 
 }
@@ -56,35 +56,23 @@ export type PokemonsPreviewDataStatus = {
   isLoading: boolean
 }
 
-export type GenPage = {
-  id: number,
-  name: string,
-  abilities: NamedAPIResource[],
-  names: Name[],
-  main_region: NamedAPIResource,
-  moves: NamedAPIResource[],
-  pokemon_species: NamedAPIResource[],
-  types: NamedAPIResource[],
-  version_groups: NamedAPIResource[]
-}
-
-type FormDescription = {
+export type FormDescription = {
   description: string,
   language: NamedAPIResource 
 }
 
-type Genera = {
+export type Genera = {
   genus: string,
   language: NamedAPIResource
 }
 
-type PalParkEncounter = {
+export type PalParkEncounter = {
   area: NamedAPIResource,
   base_score: number,
   rate: number
 }
 
-type PokedexNumber = {
+export type PokedexNumber = {
   entry_number: number,
   pokedex: NamedAPIResource
 }
@@ -94,74 +82,44 @@ export type Variety = {
   pokemon: NamedAPIResource
 }
 
-export type SpeciesPage = {
-  base_hapiness: number,
-  capture_rate: number,
-  color: NamedAPIResource,
-  egg_groups: NamedAPIResource[],
-  evolution_chain: { url: string },
-  evolves_from_species: NamedAPIResource | null,
-  flavor_text_entries: PokedexEntry[],
-  form_description: FormDescription[],
-  forms_switchable: boolean,
-  gender_rate: number,
-  genera: Genera[],
-  generation: NamedAPIResource,
-  growth_rate: NamedAPIResource,
-  habitat: NamedAPIResource,
-  has_gender_differences: boolean,
-  hatch_counter: number,
-  id: number,
-  is_baby: boolean,
-  is_legendary: boolean,
-  is_mythical: boolean,
-  name: string,
-  names: Name[],
-  order: number,
-  pal_park_encounters: PalParkEncounter[],
-  pokedex_numbers: PokedexNumber[],
-  shape: NamedAPIResource,
-  varieties: Variety[]
-}
-
-type Ability = {
+export type Ability = {
   ability: NamedAPIResource,
   is_hidden : boolean,
   slot: number
 }
 
-type Cries = {
+export type Cries = {
   latest: string,
   legacy: string | null
 }
 
-type GameIndiceVersion = {
+export type GameIndiceVersion = {
   game_index: number,
   version: NamedAPIResource
 }
 
-type VersionDetail = {
+export type VersionDetail = {
   rarity: number,
   version: NamedAPIResource
 }
 
-type HeldItem = {
+export type HeldItem = {
   item: NamedAPIResource,
   version_details: VersionDetail[]
 }
 
-type VersionGroupDetail = {
+export type VersionGroupDetail = {
   level_learned_at: number,
   move_learn_method: NamedAPIResource,
   version_group: NamedAPIResource
 }
 
-type Move = {
+export type Move = {
   move: NamedAPIResource,
   version_group_details: VersionGroupDetail[]
 }
 
-type OtherSprites = { 
+export type OtherSprites = { 
   dream_world: {
     front_default: string,
     front_female: string | null
@@ -188,7 +146,7 @@ type OtherSprites = {
   }
 }
 
-type Sprites = {
+export type Sprites = {
   back_default: string,
   back_female: string | null,
   back_shiny: string,
@@ -349,42 +307,18 @@ type Sprites = {
  }
 }
 
-type StatDetails = {
+export type StatDetails = {
   base_stat: number,
   effort: number,
   stat: NamedAPIResource
 }
 
-export type PokemonPage = {
-  abilities: Ability[]
-  base_experience: number,
-  cries: Cries,
-  forms: NamedAPIResource[],
-  game_indices: GameIndiceVersion[],
-  height: number,
-  held_items: HeldItem[]
-  id: number,
-  is_default: boolean,
-  location_area_encounters: string,
-  moves: Move[],
-  name: string,
-  order: number,
-  past_abilities: NamedAPIResource[],
-  past_types: NamedAPIResource[],
-  species: NamedAPIResource,
-  sprites: Sprites,
-  stats: StatDetails[],
-  types: PokemonType[],
-  weight: number
-}
-
-
-type GameIndiceGen = {
+export type GameIndiceGen = {
   game_index: number,
   generation: NamedAPIResource
 }
 
-type TypeRelations = {
+export type TypeRelations = {
   double_damage_from: NamedAPIResource[],
   double_damage_to: NamedAPIResource[]
   half_damage_from: NamedAPIResource[]
@@ -393,25 +327,9 @@ type TypeRelations = {
   no_damage_to: NamedAPIResource[]
 }
 
-type TypePokemon = {
+export type TypePokemon = {
   pokemon: NamedAPIResource,
   slot: number
-}
-
-export type PokemonTypePage = {
-  damage_relations: TypeRelations,
-  game_indices: GameIndiceGen[],
-  generation: NamedAPIResource,
-  id: number,
-  move_damage_class: NamedAPIResource | null,
-  moves: NamedAPIResource[],
-  name: string,
-  names: Name[],
-  past_damage_relations: {
-    damage_relations: TypeRelations,
-    generation: NamedAPIResource
-  }[],
-  pokemon: TypePokemon[]
 }
 
 export type ChainLink = {
@@ -421,7 +339,7 @@ export type ChainLink = {
   evolves_to: ChainLink[]
 }
 
-type EvolutionDetail = {
+export type EvolutionDetail = {
   item: NamedAPIResource | null,
   trigger: NamedAPIResource,
   gender: number | null,
@@ -440,42 +358,4 @@ type EvolutionDetail = {
   time_of_day: string,
   trade_species: NamedAPIResource | null,
   turn_upside_down: boolean
-}
-
-export type EvolutionChainPage = {
-  baby_trigger_item: NamedAPIResource | null,
-  chain: ChainLink
-  id: number
-}
-
-export type PokemonFormPage = {
-  form_name: string,
-  form_names: {
-    language: NamedAPIResource,
-    name: string
-  }[],
-  form_order: number,
-  id: number,
-  is_battle_only: boolean,
-  is_default: boolean,
-  is_mega: boolean,
-  name: string,
-  names: {
-    language: NamedAPIResource,
-    name: string
-  }[],
-  order: number,
-  pokemon: NamedAPIResource,
-  sprites: {
-    back_default: string | null,
-    back_female: string | null,
-    back_shiny: string | null,
-    back_shiny_female: string | null,
-    front_default: string | null,
-    front_female: string | null,
-    front_shiny: string | null,
-    front_shiny_female: string | null
-  },
-  types: PokemonType[],
-  version_group: NamedAPIResource
 }
